2025/12/20 based on InDev 0.0: improved instructions and scripts
The instructions to build and execute the OS were completely overhauled:
they explain the procedures more clearly and in detail, for both Windows and Linux.
The file is divided in sections, for setting up, testing, compiling and running on real hardware.

For testing the OS on real hardware, many scripts were added to automate the process:
they can be used for creating the binary image, writing it on physical devices and restoring them.

A "README.md" file was created to give an introduction to the project and describe its basics.
From now on, changes to the "master" branch will be traced in the "CHANGELOG.md" file.

There have also been general improvements throughout the whole code:
most of them aimed at improving the structure and correctness of comments and notes.